backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Werkzeug==3.0.0
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
